workspace/apify_frontend/requirements.txt
Flask==2.3.3
requests==2.31.0
pandas==2.2.3
openpyxl==3.1.2
python-dotenv==1.0.0
flask-cors==4.0.0
Pillow==9.5.0
